refactor: support dynamic algo module
