+++ dbterd/adapters/algos/test_relationship.py
@@ -46,7 +46,10 @@ def parse_metadata(data, **kwargs) -> Tuple[List[Table], List[Ref]]:
     logger.info(
         f"Collected {len(tables)} table(s) and {len(relationships)} relationship(s)"
     )
-    return (tables, relationships)
+    return (
+        sorted(tables, key=lambda tbl: tbl.node_name),
+        sorted(relationships, key=lambda rel: rel.name),
+    )
 
 
 def parse(
