--- dbterd/adapters/targets/constants.py
@@ -1,6 +1,0 @@
-class Strategies:
-    DBML_TEST_RELATIONSHIP = "dbml_test_relationship"
-    MERMAID_TEST_RELATIONSHIP = "mermaid_test_relationship"
-    PLANTUML_TEST_RELATIONSHIP = "plantuml_test_relationship"
-    D2_TEST_RELATIONSHIP = "d2_test_relationship"
-    GRAPHVIZ_TEST_RELATIONSHIP = "graphviz_test_relationship"
