+++ dbterd/adapters/targets/d2.py
@@ -1,6 +1,6 @@
 from typing import Tuple
 
-from dbterd.adapters.algos import test_relationship
+from dbterd.adapters import adapter
 from dbterd.types import Catalog, Manifest
 
 
@@ -28,7 +28,8 @@ def parse(manifest: Manifest, catalog: Catalog, **kwargs) -> str:
     Returns:
         str: D2 content
     """
-    tables, relationships = test_relationship.parse(
+    algo_module = adapter.load_algo(name=kwargs["algo"])
+    tables, relationships = algo_module.parse(
         manifest=manifest, catalog=catalog, **kwargs
     )
