--- dbterd/adapters/targets/default.py
@@ -1,7 +1,0 @@
-def default(**kwargs):
-    """Fallback option if the target is not supported yet
-
-    Returns:
-        Tuple[None, None]
-    """
-    return (None, None)
