+++ dbterd/adapters/targets/mermaid.py
@@ -1,7 +1,7 @@
 import re
 from typing import Optional, Tuple
 
-from dbterd.adapters.algos import test_relationship
+from dbterd.adapters import adapter
 from dbterd.types import Catalog, Manifest
 
 
@@ -79,7 +79,8 @@ def parse(manifest: Manifest, catalog: Catalog, **kwargs) -> str:
     Returns:
         str: Mermaid content
     """
-    tables, relationships = test_relationship.parse(
+    algo_module = adapter.load_algo(name=kwargs["algo"])
+    tables, relationships = algo_module.parse(
         manifest=manifest, catalog=catalog, **kwargs
     )
