+++ tests/unit/adapters/test_base.py
@@ -7,7 +7,7 @@ import pytest
 from dbterd import default
 from dbterd.adapters.base import Executor
 from dbterd.adapters.dbt_core.dbt_invocation import DbtInvocation
-from dbterd.adapters.targets.dbml import dbml_test_relationship
+from dbterd.adapters.targets import dbml
 
 
 class TestBase:
@@ -30,14 +30,14 @@ class TestBase:
     @mock.patch("dbterd.adapters.base.DbtCloudMetadata.query_erd_data")
     @mock.patch("dbterd.adapters.base.Executor._Executor__save_result")
     def test___run_metadata_by_strategy_with_not_implemented_algo(
-        self, mock_query_erd_data, mock_save_result
+        self, mock_save_result, mock_query_erd_data
     ):
-        result = Executor(
-            ctx=click.Context(command=click.BaseCommand("dummy"))
-        )._Executor__run_metadata_by_strategy(target="dbml", algo="notfound")
-        assert result is None
+        with pytest.raises(Exception):
+            Executor(
+                ctx=click.Context(command=click.BaseCommand("dummy"))
+            )._Executor__run_metadata_by_strategy(target="dbml", algo="notfound")
         mock_query_erd_data.assert_called_once()
-        mock_save_result.assert_called_once()
+        mock_save_result.call_count == 0
 
     @mock.patch("builtins.open")
     def test___save_result(self, mock_open):
@@ -246,7 +246,7 @@ class TestBase:
         ) == dict(algo="test_relationship", select=["irr"], exclude=[])
         assert mock_find_related_nodes_by_id.call_count == 1
 
-    @mock.patch("dbterd.adapters.targets.dbml.dbml_test_relationship.run")
+    @mock.patch("dbterd.adapters.targets.dbml.run")
     @mock.patch("dbterd.adapters.base.Executor._Executor__get_operation")
     @mock.patch("dbterd.adapters.base.Executor._Executor__read_manifest")
     @mock.patch("dbterd.adapters.base.Executor._Executor__read_catalog")
@@ -262,7 +262,7 @@ class TestBase:
         worker = Executor(ctx=click.Context(command=click.BaseCommand("dummy")))
 
         mock_parent = mock.Mock()
-        mock_get_operation.return_value = dbml_test_relationship.run
+        mock_get_operation.return_value = dbml.run
         mock_parent.attach_mock(mock_get_operation, "mock_get_operation")
         mock_read_catalog.return_value = dict()
         mock_parent.attach_mock(mock_read_catalog, "mock_read_catalog")
